bot/requirements.txt
opencv-contrib-python-headless
torch
torchvision
numpy
Pillow
aiogram